requirements.txt
python-daemon
bottle
